+++ utils/Vector2D.h
@@ -25,9 +25,9 @@
 
 class Vector2D {
   public:
-    Vector2D(float _x,float _y):x(_x),y(_y),updateTime(0),accessTime(0) {/*Normalize();*/}
+    Vector2D(double _x,double _y):x(_x),y(_y),updateTime(0),accessTime(0) {/*Normalize();*/}
     Vector2D():x(0),y(0),updateTime(0),accessTime(0) {}
-    void Set(float _x, float _y) { x=_x; y=_y; /*Normalize();*/}
+    void Set(double _x, double _y) { x=_x; y=_y; /*Normalize();*/}
 
     void SetUpdateTime(double t) {updateTime=t;}
     double GetUpdateTime() {return updateTime;}
@@ -45,10 +45,10 @@ class Vector2D {
     inline void operator -=(Vector2D const& other){x-=other.x;y-=other.y;}
     inline Vector2D operator -()const{return Vector2D(-x,-y);}
     // Slope angle of this vector
-    inline float atan(){ return atan2(y,x); }
+    inline double atan(){ return atan2(y,x); }
     // Square
-    inline float sq(){ return (*this) * (*this); }
-    inline float len(){ return sqrt(sq()); }
+    inline double sq(){ return (*this) * (*this); }
+    inline double len(){ return sqrt(sq()); }
 
     friend Vector2D operator /(const Vector2D& vec, const double num)
     {
@@ -73,21 +73,21 @@ class Vector2D {
     inline void operator +=(const Vector2D& v2) { x +=v2.x; y +=v2.y; }
 
     //                                                //private:
-    float x, y;
+    double x, y;
     double updateTime, accessTime;
 
     void Normalize()
     {
       if ((x==0)&&(y==0))
         return;
-      float magnitude(len());
+      double magnitude(len());
       x /= magnitude;
       y /= magnitude;
     }
 };
 
 // Calculate determinant of vector
-inline float det(Vector2D const& v1, Vector2D const& v2){
+inline double det(Vector2D const& v1, Vector2D const& v2){
   return v1.x*v2.y - v1.y*v2.x;
 }
